Rework strided buffer descriptor to be op/pattern based instead of IR walking.
This was necessary to eliminate phase ordering issues with respect to materializing microkernels and needing to do further buffer optimizations (i.e. allocation fusion, etc), which needs to add additional view transforms. By having an explicit resolution pass, this allows us to only fix certain indexing math once we have made all buffer transformations. It is also just a better separation of concerns.

We are talking about doing something like this upstream but it may need to be more general, and this is the essence of simplicity. At a minimum, it should show us that it can be done.

This also fixes a bug that wasn't showing up in current tests: it was possible to get offsets of microkernels wrong due to phase ordering of ABI creation.